docs(hub): use one script for the rest of the hub generation strategy
