+++ api/hubsaas/generation-strategy.js
@@ -1,21 +1,1 @@
-const { makeServerDynamic } = require("../make-server-dynamic");
-const removeDuplicateVersionBadge = require("../remove-duplicate-version-badge");
-const { filterByAvailability } = require("../filter-availability");
-const path = require("path");
-
-function preGenerateDocs(config) {
-  const specFilePath = config.specPath;
-  const specDir = path.dirname(specFilePath);
-
-  makeServerDynamic(specFilePath);
-  filterByAvailability(specDir, "saas");
-}
-
-function postGenerateDocs(config) {
-  removeDuplicateVersionBadge(`${config.outputDir}/hub-api.info.mdx`);
-}
-
-module.exports = {
-  preGenerateDocs,
-  postGenerateDocs,
-};
+module.exports = require("../hub-generation-strategy")("saas");
